Remove non-derivations from packages
`buildUnisonFromTranscript` and `buildUnisonShareProject` aren’t derivations and
thus shouldn’t be included in the `packages` output (`nix flake check` fails
with them there).

This moves them to the `lib` output.

It also adds an example `homeConfiguration` to test that the overlay actually
does what it says. And thus fixes the conditional issues with `darwin-security-hack`.