Move Unison library functions
They were originally like `pkgs.buildUnisonFromTranscript` and at some point I
moved them to `lib.buildUnisonFromTranscript`, but neither of those is the
correct place, the former because they’re not packages and the latter because
they depend on `pkgs`.

Unfortunately, there’s not an official flake story for this sort of thing yet,
so I made an un-official output, `packagesLib` and in the overlay they are
packaged similar to other language tooling (e.g., Haskell) –
`pkgs.unison.lib.buildFromTranscript`.